+++ frontend/src/API/user.ts
@@ -3,17 +3,42 @@ import type {
   UserProfileResponse,
   UpdateProfileRequest,
   UpdateProfileResponse,
+  UserProfileDetailsResponse,
 } from "./types";
 
 export async function fetchUserProfile(
   username: string,
-  password: string,
+
 ): Promise<UserProfileResponse> {
   const res = await fetch(
     `${BASE_URL}/api/user/profile?username=${encodeURIComponent(username)}`,
+    {
+      method: "GET",
+      credentials: "include",
+    },
   );
   if (!res.ok) throw new Error("Failed to fetch user profile");
-  return res.json();
+  const ress = await res.json();
+  console.log(ress);
+  return ress;
+}
+
+export async function fetchUserProfileDetails(
+  username: string,
+): Promise<UserProfileDetailsResponse> {
+  const res = await fetch(
+    `${BASE_URL}/api/user/profile/${encodeURIComponent(username)}`,
+    {
+      method: "GET",
+      credentials: "include",
+    },
+  );
+
+  if (!res.ok) {
+    throw new Error("Failed to fetch user profile details");
+  }
+
+  return await res.json();
 }
 
 export async function fetchUpdateProfile(
@@ -28,6 +53,7 @@ export async function fetchUpdateProfile(
   const response = await fetch(`${BASE_URL}/api/user/profile/update`, {
     method: "POST",
     body: formData,
+    credentials: "include",
   });
   if (!response.ok) {
     throw new Error(`Failed to update profile: ${response.statusText}`);